Update .travis.yml
Travis does not upgrade Docker to the latest version
(18.09) but to an older one (18.06), which does not
support all the experimental features used in NCP
Dockerfiles. Therefore, we install it with the convenience
script in order to get Docker 18.09. Git push to figure
out if the errors were deriving from the older docker
version.